+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #
 - Add missing `ContainerSpec` fields, use correct `TaskSpec` in `ServiceSpec`
 - Fix `ContainerConfig::exposed_ports` serialization
+- Fix `ContainerConfig` desserialization (`cmd` field might be ommited from the response)
 
 # 0.4.0
 - Fix list endpoints